revert trufflehog v3
