+++ backend/engine/plugins/truffle_hog/settings.json
@@ -1,5 +1,5 @@
 {
-    "name": "Trufflehog (Legacy) Scanner",
+    "name": "TruffleHog Scanner",
     "type": "secrets",
     "image": "$ECR/artemis/python:3-latest"
 }
