+++ backend/lambdas/api/repo/tests/test_parse_event.py
@@ -227,7 +227,6 @@ EXPECTED_EVENT_RESULT_1 = {
         "-base_images",
         "-node_dependencies",
         "-truffle_hog",
-        "-trufflehog",
         "-bundler_audit",
         "-php_sensio_security_checker",
         "-github_repo_health",
