+++ backend/lambdas/api/spec.yaml
@@ -2079,7 +2079,6 @@ components:
               - python_code_checker
               - technology_discovery
               - truffle_hog
-              - trufflehog
               - brakeman
               - bundler_audit
               - bandit
