+++ ui/src/api/__tests__/client.test.ts
@@ -290,7 +290,7 @@ describe("api client", () => {
 				expect.objectContaining({
 					data: expect.objectContaining({
 						categories: categories,
-						plugins: ["-ghas_secrets", "-truffle_hog", "-trufflehog"],
+						plugins: ["-ghas_secrets", "-truffle_hog"],
 					}),
 					url: `${data.vcsOrg}/${data.repo}`,
 				})
